chore(integration/kafka): split instrumentation into producer and consumer submodules
The kafka-go instrumentation lived in a single module that hooked both
(*Writer).WriteMessages and (*Reader).ReadMessage. grpc and net/http instead
split into client/ and server/ submodules so users can enable only one side.

Follow that convention for kafka: move the producer hooks into producer/ and
the consumer hooks into consumer/, each with its own go.mod and otelc.yaml.
The shared semconv package stays under the root instrumentation module, as it
does for grpc and net/http. With otel.instrumentation.go, a user can now pull
in just the producer or just the consumer instrumentation.

Addresses #635.
