--- instrumentation/github.com/segmentio/kafka-go/otelc.yaml
@@ -1,21 +1,0 @@
-kafka_writer_writemessages:
-  target: github.com/segmentio/kafka-go
-  where:
-    func: WriteMessages
-    recv: "*Writer"
-  do:
-    - inject_hooks:
-        before: BeforeWriteMessages
-        after: AfterWriteMessages
-        path: "go.opentelemetry.io/otelc/instrumentation/github.com/segmentio/kafka-go"
-
-kafka_reader_readmessage:
-  target: github.com/segmentio/kafka-go
-  where:
-    func: ReadMessage
-    recv: "*Reader"
-  do:
-    - inject_hooks:
-        before: BeforeReadMessage
-        after: AfterReadMessage
-        path: "go.opentelemetry.io/otelc/instrumentation/github.com/segmentio/kafka-go"
