+++ instrumentation/github.com/segmentio/kafka-go/producer/go.mod
@@ -1,6 +1,12 @@
-module go.opentelemetry.io/otelc/instrumentation/github.com/segmentio/kafka-go
+module go.opentelemetry.io/otelc/instrumentation/github.com/segmentio/kafka-go/producer
 
 go 1.25.0
+
+replace go.opentelemetry.io/otelc/pkg => ../../../../../pkg
+
+replace go.opentelemetry.io/otelc/pkg/runtime => ../../../../../pkg/runtime
+
+replace go.opentelemetry.io/otelc/instrumentation => ../../../..
 
 require (
 	github.com/segmentio/kafka-go v0.4.51
@@ -8,6 +14,7 @@ require (
 	go.opentelemetry.io/otel v1.44.0
 	go.opentelemetry.io/otel/sdk v1.44.0
 	go.opentelemetry.io/otel/trace v1.44.0
+	go.opentelemetry.io/otelc/instrumentation v0.0.0-00010101000000-000000000000
 	go.opentelemetry.io/otelc/pkg v0.0.0-00010101000000-000000000000
 	go.opentelemetry.io/otelc/pkg/runtime v0.0.0-00010101000000-000000000000
 )
@@ -56,11 +63,7 @@ require (
 	golang.org/x/text v0.37.0 // indirect
 	google.golang.org/genproto/googleapis/api v0.0.0-20260526163538-3dc84a4a5aaa // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20260526163538-3dc84a4a5aaa // indirect
-	google.golang.org/grpc v1.81.1 // indirect
+	google.golang.org/grpc v1.82.0 // indirect
 	google.golang.org/protobuf v1.36.11 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
-
-replace go.opentelemetry.io/otelc/pkg => ../../../../pkg
-
-replace go.opentelemetry.io/otelc/pkg/runtime => ../../../../pkg/runtime
